small changes & routing
