+++ src/app/auth/login/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { AuthLogin } from "@/components/AuthLogin/AuthLogin";
+import { AuthLogin } from "@/components/auth/AuthLogin";
 
 export default function Login() {
   return <AuthLogin />;
