--- src/components/AuthLogin/UabcHeaderText.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export const UabcHeaderText = () => {
-  return (
-    <div className="text-center">
-      <span className="font-proxima text-[4rem] font-bold leading-[5rem] text-primary dark:text-white tracking-tight">
-        UABC
-      </span>
-    </div>
-  );
-};
