+++ src/components/UabcHeaderText.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const UabcHeaderText = () => (
+  <div className="text-center">
+    <span className="font-proxima text-[4rem] font-bold leading-[5rem] text-primary dark:text-white tracking-tight">
+      UABC
+    </span>
+  </div>
+);
